Search_Tutorial/requirements.txt
python-dotenv>=1.0.0
pandas>=1.5.0 